# common-platform-user-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-user-mgmt
  description: User Management Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/users
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/users/{userId}:
    get:
      tags:
        - User Accounts
      summary: Get user account by ID
      description: Returns a user account by its ID
      operationId: getUserAccountById
      parameters:
        - name: userId
          in: path
          description: ID of the user account to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved user account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
        '404':
          description: User account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
    put:
      tags:
        - User Accounts
      summary: Update an existing user account
      description: Updates an existing user account and returns the updated user account
      operationId: updateUserAccount
      parameters:
        - name: userId
          in: path
          description: ID of the user account to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAccountDTO'
        required: true
      responses:
        '200':
          description: User account successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
        '404':
          description: User account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
    delete:
      tags:
        - User Accounts
      summary: Delete a user account
      description: Deletes a user account by its ID
      operationId: deleteUserAccount
      parameters:
        - name: userId
          in: path
          description: ID of the user account to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User account successfully deleted
        '404':
          description: User account not found
...
  /api/v1/users/filter:
    post:
      tags:
        - User Accounts
      summary: Get all user accounts with filtering
      description: Returns a paginated list of user accounts based on filter criteria
      operationId: filterUserAccounts
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestUserAccountDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved user accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
  /api/v1/users:
    post:
      tags:
        - User Accounts
      summary: Create a new user account
      description: Creates a new user account and returns the created user account
      operationId: createUserAccount
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAccountDTO'
        required: true
      responses:
        '201':
          description: User account successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccountDTO'
...
  /api/v1/audit-logs/users/{userId}:
    get:
      tags:
        - Audit Logs
      summary: Get audit logs for a user
      description: Returns all audit logs for a specific user
      operationId: getAuditLogsByUserId
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
...
components:
  schemas:
    UserAccountDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        fullName:
          type: string
        nickname:
          type: string
        email:
          type: string
        userType:
          type: string
          enum:
            - EMPLOYEE
            - DISTRIBUTOR
        branchId:
          type: string
          format: uuid
        distributorId:
          type: string
          format: uuid
        departmentId:
          type: string
          format: uuid
        positionId:
          type: string
          format: uuid
        jobTitle:
          type: string
        avatarUrl:
          type: string
        themePreference:
          type: string
          enum:
            - LIGHT
            - DARK
            - SYSTEM
        languagePreference:
          type: string
        locale:
          type: string
        timezone:
          type: string
        contactPhone:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
    FilterRequestUserAccountDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/UserAccountDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...